changed FetchData to IncidentCreationForm | moved constants interfaces to single file
