--- constants.js
@@ -1,1 +1,0 @@
-export const NEXT_PUBLIC_AUTH_TOKEN="c12226f5-9156-47cd-bf7c-a36379785a25"
